:art: chore: 重构设置文件
