+++ src/app/settings/Settings/touch/ActionList/Actions/Delete.tsx
