+++ src/app/settings/Settings/touch/ActionList/List/MaleList.tsx
@@ -2,11 +2,11 @@ import { isEqual } from 'lodash-es';
 import React, { memo } from 'react';
 import { useTranslation } from 'react-i18next';
 
+import Header from '@/components/Header';
 import { configSelectors, useSettingStore } from '@/store/setting';
 import { GenderEnum } from '@/types/agent';
 import { TouchAreaEnum } from '@/types/touch';
 
-import Header from '../../components/Header';
 import AddOrEdit from '../Actions/AddOrEdit';
 import ListItem from '../ListItem';
 
@@ -16,8 +16,8 @@ interface AreaListProps {
 
 const AreaList = memo((props: AreaListProps) => {
   const { currentTouchArea } = props;
-  const female = useSettingStore(
-    (s) => configSelectors.getTouchActionsByGenderAndArea(s, GenderEnum.FEMALE, currentTouchArea),
+  const touchList = useSettingStore(
+    (s) => configSelectors.getTouchActionsByGenderAndArea(s, GenderEnum.MALE, currentTouchArea),
     isEqual,
   );
   const { t } = useTranslation('features');
@@ -25,17 +25,17 @@ const AreaList = memo((props: AreaListProps) => {
   return (
     <>
       <Header
-        title={t('agent.female')}
-        extra={<AddOrEdit isEdit={false} touchArea={currentTouchArea} gender={GenderEnum.FEMALE} />}
+        title={t('agent.male')}
+        extra={<AddOrEdit isEdit={false} touchArea={currentTouchArea} gender={GenderEnum.MALE} />}
       />
-      {female.map((item, index) => {
+      {touchList.map((item, index) => {
         return (
           <ListItem
             item={item}
             currentTouchArea={currentTouchArea}
             index={index}
             key={index}
-            gender={GenderEnum.FEMALE}
+            gender={GenderEnum.MALE}
           />
         );
       })}
