+++ src/app/settings/Settings/touch/ActionList/List/OtherList.tsx
@@ -2,11 +2,11 @@ import { isEqual } from 'lodash-es';
 import React, { memo } from 'react';
 import { useTranslation } from 'react-i18next';
 
+import Header from '@/components/Header';
 import { configSelectors, useSettingStore } from '@/store/setting';
 import { GenderEnum } from '@/types/agent';
 import { TouchAreaEnum } from '@/types/touch';
 
-import Header from '../../components/Header';
 import AddOrEdit from '../Actions/AddOrEdit';
 import ListItem from '../ListItem';
 
@@ -17,7 +17,7 @@ interface AreaListProps {
 const AreaList = memo((props: AreaListProps) => {
   const { currentTouchArea } = props;
   const touchList = useSettingStore(
-    (s) => configSelectors.getTouchActionsByGenderAndArea(s, GenderEnum.MALE, currentTouchArea),
+    (s) => configSelectors.getTouchActionsByGenderAndArea(s, GenderEnum.OTHER, currentTouchArea),
     isEqual,
   );
   const { t } = useTranslation('features');
@@ -25,8 +25,8 @@ const AreaList = memo((props: AreaListProps) => {
   return (
     <>
       <Header
-        title={t('agent.male')}
-        extra={<AddOrEdit isEdit={false} touchArea={currentTouchArea} gender={GenderEnum.MALE} />}
+        title={t('agent.other')}
+        extra={<AddOrEdit isEdit={false} touchArea={currentTouchArea} gender={GenderEnum.OTHER} />}
       />
       {touchList.map((item, index) => {
         return (
@@ -35,7 +35,7 @@ const AreaList = memo((props: AreaListProps) => {
             currentTouchArea={currentTouchArea}
             index={index}
             key={index}
-            gender={GenderEnum.MALE}
+            gender={GenderEnum.OTHER}
           />
         );
       })}
